refactor(glyph): split oversized pug oracle sources and sharpen evidence
